requirements.txt
pandas
numpy
torch
transformers
sentencepiece
datasets
matplotlib
python-dotenv
nltk
spacy